requirements.txt
selenium
requests
python-dotenv
chromedriver-autoinstaller
flask
webdriver_manager